move out petition file renderer to its own component
